Rename dateWrapper to dateUtils.
