+++ src/client/js/components/MeasureTime.js
@@ -1,5 +1,5 @@
 import * as timeUtils from 'scripts/utils/timeConversion';
-import { now } from '../utils/dateWrapper';
+import { now } from 'client/js/utils/dateUtils';
 
 var MeasureTime = function (day = timeUtils.asDay(now()), minutes = 0) {
     this.day = day;
