+++ src/client/js/components/ModelHandler.js
@@ -1,7 +1,7 @@
 import * as MeasureTime from './MeasureTime';
 import * as timeUtils from 'scripts/utils/timeConversion';
 import * as store from './PersistentStore';
-import { now } from '../utils/dateWrapper';
+import { now } from 'client/js/utils/dateUtils';
 
 const storeConfig = {
     requiredDbVersion: 2,
